+++ examples/snippets/meter_event_stream.ts
@@ -1,3 +1,15 @@
+/**
+ * meter_event_stream.ts - Use the high-throughput meter event stream to report create billing meter events.
+ *
+ * In this example, we:
+ *   - create a meter event session and store the session's authentication token
+ *   - define an event with a payload
+ *   - use the meterEventStream service to create an event stream that reports this event
+ *
+ * This example expects a billing meter with an event_name of 'alpaca_ai_tokens'.  If you have
+ * a different meter event name, you can change it before running this example.
+ */
+
 import {Stripe} from 'stripe';
 
 const apiKey = '{{API_KEY}}';