+++ types/Accounts.d.ts
@@ -335,6 +335,11 @@ declare module 'stripe' {
         jp_bank_transfer_payments?: Capabilities.JpBankTransferPayments;
 
         /**
+         * The status of the KakaoPay capability of the account, or whether the account can directly process KakaoPay payments.
+         */
+        kakao_pay_payments?: Capabilities.KakaoPayPayments;
+
+        /**
          * The status of the Klarna payments capability of the account, or whether the account can directly process Klarna charges.
          */
         klarna_payments?: Capabilities.KlarnaPayments;
@@ -343,6 +348,11 @@ declare module 'stripe' {
          * The status of the konbini payments capability of the account, or whether the account can directly process konbini charges.
          */
         konbini_payments?: Capabilities.KonbiniPayments;
+
+        /**
+         * The status of the KrCard capability of the account, or whether the account can directly process KrCard payments.
+         */
+        kr_card_payments?: Capabilities.KrCardPayments;
 
         /**
          * The status of the legacy payments capability of the account.
@@ -375,6 +385,11 @@ declare module 'stripe' {
         mx_bank_transfer_payments?: Capabilities.MxBankTransferPayments;
 
         /**
+         * The status of the NaverPay capability of the account, or whether the account can directly process NaverPay payments.
+         */
+        naver_pay_payments?: Capabilities.NaverPayPayments;
+
+        /**
          * The status of the OXXO payments capability of the account, or whether the account can directly process OXXO charges.
          */
         oxxo_payments?: Capabilities.OxxoPayments;
@@ -383,6 +398,11 @@ declare module 'stripe' {
          * The status of the P24 payments capability of the account, or whether the account can directly process P24 charges.
          */
         p24_payments?: Capabilities.P24Payments;
+
+        /**
+         * The status of the Payco capability of the account, or whether the account can directly process Payco payments.
+         */
+        payco_payments?: Capabilities.PaycoPayments;
 
         /**
          * The status of the paynow payments capability of the account, or whether the account can directly process paynow charges.
@@ -413,6 +433,11 @@ declare module 'stripe' {
          * The status of the RevolutPay capability of the account, or whether the account can directly process RevolutPay payments.
          */
         revolut_pay_payments?: Capabilities.RevolutPayPayments;
+
+        /**
+         * The status of the SamsungPay capability of the account, or whether the account can directly process SamsungPay payments.
+         */
+        samsung_pay_payments?: Capabilities.SamsungPayPayments;
 
         /**
          * The status of the SEPA customer_balance payments (EUR currency) capability of the account, or whether the account can directly process SEPA customer_balance charges.
@@ -453,6 +478,21 @@ declare module 'stripe' {
          * The status of the banking capability, or whether the account can have bank accounts.
          */
         treasury?: Capabilities.Treasury;
+
+        /**
+         * The status of the treasury_evolve capability of the account.
+         */
+        treasury_evolve?: Capabilities.TreasuryEvolve;
+
+        /**
+         * The status of the treasury_fifth_third capability of the account.
+         */
+        treasury_fifth_third?: Capabilities.TreasuryFifthThird;
+
+        /**
+         * The status of the treasury_goldman_sachs capability of the account.
+         */
+        treasury_goldman_sachs?: Capabilities.TreasuryGoldmanSachs;
 
         /**
          * The status of the TWINT capability of the account, or whether the account can directly process TWINT charges.
@@ -522,9 +562,13 @@ declare module 'stripe' {
 
         type JpBankTransferPayments = 'active' | 'inactive' | 'pending';
 
+        type KakaoPayPayments = 'active' | 'inactive' | 'pending';
+
         type KlarnaPayments = 'active' | 'inactive' | 'pending';
 
         type KonbiniPayments = 'active' | 'inactive' | 'pending';
+
+        type KrCardPayments = 'active' | 'inactive' | 'pending';
 
         type LegacyPayments = 'active' | 'inactive' | 'pending';
 
@@ -538,9 +582,13 @@ declare module 'stripe' {
 
         type MxBankTransferPayments = 'active' | 'inactive' | 'pending';
 
+        type NaverPayPayments = 'active' | 'inactive' | 'pending';
+
         type OxxoPayments = 'active' | 'inactive' | 'pending';
 
         type P24Payments = 'active' | 'inactive' | 'pending';
+
+        type PaycoPayments = 'active' | 'inactive' | 'pending';
 
         type PaynowPayments = 'active' | 'inactive' | 'pending';
 
@@ -553,6 +601,8 @@ declare module 'stripe' {
         type RechnungPayments = 'active' | 'inactive' | 'pending';
 
         type RevolutPayPayments = 'active' | 'inactive' | 'pending';
+
+        type SamsungPayPayments = 'active' | 'inactive' | 'pending';
 
         type SepaBankTransferPayments = 'active' | 'inactive' | 'pending';
 
@@ -569,6 +619,12 @@ declare module 'stripe' {
         type Transfers = 'active' | 'inactive' | 'pending';
 
         type Treasury = 'active' | 'inactive' | 'pending';
+
+        type TreasuryEvolve = 'active' | 'inactive' | 'pending';
+
+        type TreasuryFifthThird = 'active' | 'inactive' | 'pending';
+
+        type TreasuryGoldmanSachs = 'active' | 'inactive' | 'pending';
 
         type TwintPayments = 'active' | 'inactive' | 'pending';