+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 17.2.0 - 2024-10-09
+* [#2201](https://github.com/stripe/stripe-node/pull/2201) Add fetchRelatedObject to V2 Events if needed
+  * `fetchRelatedObject` is added to events retrieved using `stripe.v2.core.events` and can be used to easily fetch the Stripe object related to a retrieved event
+
 ## 17.2.0-beta.2 - 2024-10-08
 * [#2180](https://github.com/stripe/stripe-node/pull/2180) Update generated code for beta
   * Add support for `submit_card` test helper method on resource `Issuing.Card`